Resnet Layers' optimization are ongoing
