+++ proj/solution1/.autopilot/db/all.directive.json
@@ -1,7 +1,7 @@
 [
   {
     "functionLabel": "",
-    "functionName": "resnet_top_4",
+    "functionName": "resnet_top_1",
     "id": 0,
     "ifcond": "",
     "insert_position": "",
@@ -11,13 +11,13 @@
       "option": [
         {
           "name": "name",
-          "value": "resnet_top_4"
+          "value": "resnet_top_1"
         }
       ]
     },
     "slx": false,
-    "sourceFile": "/nethome/mjung76/resnet_rpn_fpn/script_resnet4.tcl",
-    "sourceLine": 27,
+    "sourceFile": "/nethome/mjung76/resnet_rpn_fpn/script_resnet1.tcl",
+    "sourceLine": 28,
     "success": true
   }
 ]
