+++ proj/solution1/solution1.directive
@@ -1,8 +1,8 @@
 <AutoPilot:ProjectResources xmlns:AutoPilot="com.autoesl.autopilot.ui.directiveView">
     <sourceFile name="">
-        <directive functionName="resnet_top_4" label="" functionLabel="">
+        <directive functionName="resnet_top_1" label="" functionLabel="">
             <pragma name="TOP">
-                <option name="name" value="resnet_top_4"/>
+                <option name="name" value="resnet_top_1"/>
             </pragma>
         </directive>
     </sourceFile>
