+++ resnet_layer1.cpp
@@ -1,5 +1,5 @@
-#include "qdtrack.h"
-#include "resnet_util.h"
+#include "qdtrack_resnet1.h"
+#include "resnet_util1.h"
 
 void resnet_layer1(
         fm_t   resnet_layer1_input_fm[RESNET_LAYER1_0_CONV1_IN_CH][RESNET_LAYER1_0_FM_HEIGHT][RESNET_LAYER1_0_FM_WIDTH],
