--- resnet_layers.cpp
@@ -1,7 +1,0 @@
-#include "resnet_util.h"
-
-using namespace std;
-
-#include "resnet_layer0.cpp"
-#include "resnet_layer1.cpp"
-// #include "resnet_layer2.cpp"
